+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Use mitogen for handler io
 - Add default null value to file owner and group
+- Add files::path_join plugin
 
 ## v0.6.1 - 2024-07-22